kp: revert sus_mount to v1.5.5
https://github.com/sidex15/android_kernel_lge_sm8150/commit/1acc546c34254774b6f6d67a412e3d115fd6d739

SUSFS v1.5.6-v1.5.7 introduced mount ID reordering optimizations. But since it's for GKI or a specific fix for some devices, in this device it could
cause more harm than good which it leaks mount-related detections.

For now revert the sus_mount mechanism to v1.5.5 as it's still okay and stable.

Co-authored-by: cvnertnc <the.civanertunc@gmail.com>
Co-authored-by: sidex15 <24408329+sidex15@users.noreply.github.com>